services/service_b/requirements.txt
apturl==0.5.2
argcomplete==1.8.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
Flask==1.1.2
future==0.18.2
grpcio==1.33.1
grpcio-tools==1.33.1
httplib2==0.14.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
progressbar==2.5
protobuf==3.13.0
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.14.0
python-gnupg==0.4.5
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
systemd-python==234
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
Ubuntu-Make==18.9
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
Werkzeug==1.0.1
xdiagnose==3.8.9
xkit==0.0.0
